+++ src/Modulos/categorias/dto/categoria.dto.ts
@@ -1,0 +1,8 @@
+import { IsString } from "class-validator";
+
+export class CategoriaDTO{
+    @IsString()
+    nombre: string;
+    @IsString()
+    readonly descripcion: string;
+}
