+++ src/Modulos/categorias/interfaces/categoria.interface.ts
