+++ src/Modulos/recetas/interfaces/receta.interface.ts
