+++ src/Modulos/roles/dto/rol.dto.ts
@@ -1,0 +1,12 @@
+import { IsString } from "class-validator";
+
+export class RolDTO {
+    @IsString()
+    readonly _id?: string;
+
+    @IsString()
+    nombre: string;   
+    
+    @IsString()
+    readonly descripcion: string;
+}
