+++ src/Modulos/roles/interfaces/rol.interface.ts
