+++ src/Modulos/roles/schemas/Rol.json
