+++ src/Modulos/users/dto/usuarioEditado.dto.ts
@@ -1,0 +1,28 @@
+import { IsDate, IsEmail, IsOptional, IsString, MinLength } from "class-validator";
+export class UsuarioEditadoDTO {
+    @IsString()
+    @MinLength(3)
+    @IsOptional()
+    usuario: string;
+    
+    @IsString()
+    @MinLength(3)
+    @IsOptional()
+    password: string;
+
+    @IsEmail()
+    @IsOptional()
+    email: string;
+
+    @IsString()
+    @IsOptional()
+    rol?: string;
+    
+    @IsDate()
+    @IsOptional()
+    readonly fechaCreacion?: Date;
+
+    @IsDate()
+    @IsOptional()
+    readonly fechaActualizacion?: Date;
+}
