+++ src/Modulos/users/interfaces/usuarios.interface.ts
@@ -1,0 +1,10 @@
+
+
+export interface UsuarioInterface {
+    readonly usuario: string;
+    readonly password: string;
+    readonly email: string;
+    readonly rol?: string;
+    readonly fechaCreacion?: Date;
+    fechaActualizacion?: Date;
+}
