+++ src/auth/const/jwt.constant.ts
@@ -1,0 +1,1 @@
+export const JWT_SECRET = process.env.JWT_SECRET || 'secret';
