--- src/categorias/dto/categoria.dto.ts
@@ -1,4 +1,0 @@
-export class CategoriaDTO{
-    readonly nombre: string;
-    readonly descripcion: string;
-}
