+++ src/decorators/roles.decorator.ts
@@ -1,0 +1,4 @@
+import { SetMetadata } from "@nestjs/common";
+
+export const ROLES_KEY = 'roles';
+export const Rol = (rol: string) => SetMetadata(ROLES_KEY, rol);
