--- src/roles/dto/rol.dto.ts
@@ -1,5 +1,0 @@
-export class RolDTO {
-    readonly _id?: string;
-    nombre: string;     
-    readonly descripcion: string;
-}
