--- src/users/interfaces/usuarios.interface.ts
@@ -1,10 +1,0 @@
-import { RolDocument } from "src/roles/schemas/rol.schema";
-
-export interface UsuarioInterface {
-    readonly usuario: string;
-    readonly password: string;
-    readonly email: string;
-    readonly rol: string;
-    readonly fechaCreacion: Date;
-    fechaActualizacion: Date;
-}
